3-2 refactor joined TestStudentController & TestProctorController to TestRoomController
