+++ backend/testhelper/src/main/java/kr/ac/ajou/da/testhelper/test/room/dto/GetTestStudentRoomResDto.java
@@ -1,16 +1,13 @@
-package kr.ac.ajou.da.testhelper.test.proctor.dto;
+package kr.ac.ajou.da.testhelper.test.room.dto;
 
 import com.amazonaws.services.securitytoken.model.Credentials;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
-import java.util.List;
-
 @RequiredArgsConstructor
 @Getter
-public class PostTestStudentRoomResDto {
-
+public class GetTestStudentRoomResDto {
     private final Credentials credentials;
-    private final List<StudentRoomDto> students;
+    private final RoomDto room;
 
 }
