+++ backend/testhelper/src/main/java/kr/ac/ajou/da/testhelper/test/room/dto/PostTestStudentRoomResDto.java
@@ -1,13 +1,16 @@
-package kr.ac.ajou.da.testhelper.test.student.dto;
+package kr.ac.ajou.da.testhelper.test.room.dto;
 
 import com.amazonaws.services.securitytoken.model.Credentials;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
+import java.util.List;
+
 @RequiredArgsConstructor
 @Getter
-public class GetTestStudentRoomResDto {
+public class PostTestStudentRoomResDto {
+
     private final Credentials credentials;
-    private final RoomDto room;
+    private final List<StudentRoomDto> students;
 
 }
